pyproject.toml
[tool.poetry]
name = "mcp-knowledge-base"
version = "0.1.0"
description = "MCP client-server example"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
mcp = {extras = ["cli"], version = "1.6.0"}
openai = "1.75.0"
python-dotenv = "^1.0.0"
ipykernel = "^6.25.0"
httpx = "^0.25.0"
nest-asyncio = "^1.5.7"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false